+++ src/main/java/com/mageddo/dnsproxyserver/App.java
@@ -1,7 +1,7 @@
 package com.mageddo.dnsproxyserver;
 
+import com.mageddo.dnsproxyserver.config.Configs;
 import com.mageddo.dnsproxyserver.di.Context;
-import com.mageddo.dnsproxyserver.quarkus.QuarkusConfig;
 
 public class App {
   public static void main(String[] args) {
@@ -13,7 +13,7 @@ public class App {
     context.start();
 
     // fixme ajustar logs do logback
-    QuarkusConfig.setup(config);
+    AppSettings.setup(Configs.getInstance());
 
     // todo install as service
