+++ src/main/java/com/mageddo/dnsproxyserver/quarkus/QuarkusConfig.java
@@ -39,20 +39,4 @@ public class QuarkusConfig {
     };
   }
 
-  public static void setup(Config config) {
-
-    if (config.getLogLevel() != null) {
-      LogbackUtils.changeLogLevel("com.mageddo", config.getLogLevel().toLogbackLevel());
-    }
-
-    final var logFile = Configs.parseLogFile(config.getLogFile());
-    if (logFile == null) {
-      System.setProperty("quarkus.log.console.enable", "false");
-      System.setProperty("quarkus.log.file.enable", "false");
-    } else if (!equalsIgnoreCase(logFile, "console")) {
-      System.setProperty("quarkus.log.console.enable", "false");
-      System.setProperty("quarkus.log.file.enable", "true");
-      System.setProperty("quarkus.log.file.path", config.getLogFile());
-    }
-  }
 }
